+++ packages/components-vue/lib/vue-component-lib/utils.d.ts
@@ -1,0 +1,4 @@
+export interface InputProps<T> {
+    modelValue?: T;
+}
+export declare const defineContainer: <Props, VModelType = string | number | boolean>(name: string, defineCustomElement: any, componentProps?: string[], modelProp?: string, modelUpdateEvent?: string, externalModelUpdateEvent?: string) => (props: Props & InputProps<VModelType> & {}) => any;