+++ packages/components/src/components/slider/slider.e2e.tsx
@@ -1,19 +1,19 @@
 import { newE2EPage } from '@stencil/core/testing';
 
-describe('ifx-range', () => {
+describe('ifx-slider', () => {
     it('renders', async () => {
         const page = await newE2EPage();
-        await page.setContent('<ifx-range></ifx-range>');
+        await page.setContent('<ifx-slider></ifx-slider>');
 
-        const element = await page.find('ifx-range');
+        const element = await page.find('ifx-slider');
         expect(element).toHaveClass('hydrated');
     });
 
     it('renders changes to the value data', async () => {
         const page = await newE2EPage();
-        await page.setContent('<ifx-range value="50"></ifx-range>');
+        await page.setContent('<ifx-slider value="50"></ifx-slider>');
 
-        const element = await page.find('ifx-range >>> input');
+        const element = await page.find('ifx-slider >>> input');
 
         const value = await element.getProperty('value');
         expect(value).toEqual('50');
@@ -21,9 +21,9 @@ describe('ifx-range', () => {
 
     it('handles input change', async () => {
         const page = await newE2EPage();
-        await page.setContent('<ifx-range></ifx-range>');
+        await page.setContent('<ifx-slider></ifx-slider>');
 
-        const element = await page.find('ifx-range >>> input');
+        const element = await page.find('ifx-slider >>> input');
         await element.press('ArrowRight');
 
         const value = await element.getProperty('value');
@@ -32,42 +32,42 @@ describe('ifx-range', () => {
 
     it('disables when the disabled property is set', async () => {
         const page = await newE2EPage();
-        await page.setContent('<ifx-range disabled></ifx-range>');
+        await page.setContent('<ifx-slider disabled></ifx-slider>');
 
-        const element = await page.find('ifx-range >>> input');
+        const element = await page.find('ifx-slider >>> input');
         expect(element.getAttribute('disabled')).not.toBeNull();
     });
 
     it('renders left text', async () => {
         const page = await newE2EPage();
-        await page.setContent('<ifx-range left-text="Left Text"></ifx-range>');
+        await page.setContent('<ifx-slider left-text="Left Text"></ifx-slider>');
 
-        const element = await page.find('ifx-range >>> .left-text');
+        const element = await page.find('ifx-slider >>> .left-text');
         expect(element.textContent).toEqual('Left Text');
     });
 
     it('renders right text', async () => {
         const page = await newE2EPage();
-        await page.setContent('<ifx-range right-text="Right Text"></ifx-range>');
+        await page.setContent('<ifx-slider right-text="Right Text"></ifx-slider>');
 
-        const element = await page.find('ifx-range >>> .right-text');
+        const element = await page.find('ifx-slider >>> .right-text');
         expect(element.textContent).toEqual('Right Text');
     });
 
     it('renders percentage if showPercentage is true', async () => {
         const page = await newE2EPage();
-        await page.setContent('<ifx-range value="30" show-percentage></ifx-range>');
+        await page.setContent('<ifx-slider value="30" show-percentage></ifx-slider>');
 
-        const element = await page.find('ifx-range >>> .percentage-display');
+        const element = await page.find('ifx-slider >>> .percentage-display');
         expect(element.textContent).toEqual('30%');
     });
 
     it('emits the ifxChanged event when the input changes', async () => {
         const page = await newE2EPage();
-        await page.setContent('<ifx-range value="50"></ifx-range>');
+        await page.setContent('<ifx-slider value="50"></ifx-slider>');
         const spy = await page.spyOnEvent('ifxChanged');
 
-        const input = await page.find('ifx-range >>> input');
+        const input = await page.find('ifx-slider >>> input');
         await input.press('ArrowRight');
 
         expect(spy).toHaveReceivedEventDetail(51);